+++ src/web/FloraWeb/Templates/Pages/Home.hs
@@ -22,10 +22,10 @@ import Flora.Environment
 import FloraWeb.Templates.Types
 
 show :: FloraHTML
-show =
-  main_ $! do
-    banner
-    searchBar
+show = do
+  banner
+  searchBar
+  buttons
 
 about :: FloraHTML
 about = do
@@ -96,3 +96,15 @@ searchBar =
       button_ [type_ "submit"] $
         svg_ [xmlns_ "http://www.w3.org/2000/svg", style_ "color: gray", fill_ "none", viewBox_ "0 0 24 24", stroke_ "currentColor"] $
           path_ [stroke_linecap_ "round", stroke_linejoin_ "round", stroke_width_ "2", d_ "M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"]
+
+buttons :: FloraHTML
+buttons =
+  section_ [id_ "main-page-buttons"] $ do
+    a_ [class_ "button", href_ "https://www.haskell.org/ghcup/"] $ do
+      h2_
+        [class_ "category-card__name"]
+        "Install Haskell"
+    a_ [class_ "button", href_ "https://cabal.readthedocs.io/en/stable/intro.html"] $ do
+      h2_
+        [class_ "category-card__name"]
+        "Start with Cabal"