+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 1.0.10 -- XXXX-XX-XX
+* Add buttons to the main page for ghcup and cabal guides ([#341](https://github.com/flora-pm/flora-server/pull/341))
 * Split the project into internal libraries ([#337](https://github.com/flora-pm/flora-server/pull/337)) 
 
 ## 1.0.9 -- 2023-01-06